updated for impersonatedAuthClient if IMPERSONATED_PRINCIPAL is exported
