+++ .github/auto-approve.yml
@@ -1,12 +1,3 @@
-rules:
-- author: "release-please[bot]"
-  title: "^chore: release"
-  changedFiles:
-  - "package\\.json$"
-  - "CHANGELOG\\.md$"
-  maxFiles: 3
-- author: "renovate-bot"
-  title: "^(fix|chore)\\(deps\\):"
-  changedFiles: 
-  - "package\\.json$"
-  maxFiles: 2
+processes:
+  - "NodeDependency"
+  - "OwlBotTemplateChanges"
