+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@google-cloud/retail",
-  "version": "1.8.0",
+  "version": "1.8.1",
   "description": "Retail client for Node.js",
   "repository": "googleapis/nodejs-retail",
   "license": "Apache-2.0",
@@ -57,7 +57,7 @@
     "mocha": "^8.2.1",
     "null-loader": "^4.0.1",
     "pack-n-play": "^1.0.0-2",
-    "sinon": "^13.0.0",
+    "sinon": "^14.0.0",
     "ts-loader": "^9.0.0",
     "typescript": "^4.1.2",
     "webpack": "^5.9.0",
