+++ samples/interactive-tutorials/resources/user_events.json
@@ -1,4 +1,4 @@
-{"eventType":"home-page-view","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-12T10:27:42+00:00"}
-{"eventType":"search","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-12T10:27:42+00:00","searchQuery":"RockerJeans teenagers blue jeans"}
-{"eventType":"search","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-12T10:27:42+00:00","searchQuery":"SocksUnlimited teenagers black socks"}
-{"eventType":"detail-page-view","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-12T10:27:42+00:00","productDetails":{"product":{"id":"GGCOGAEC100616"},"quantity":3}}
+{"eventType":"home-page-view","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-10T10:27:42.000Z"}
+{"eventType":"search","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-10T10:27:42.000Z","searchQuery":"RockerJeans teenagers blue jeans"}
+{"eventType":"search","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-10T10:27:42.000Z","searchQuery":"SocksUnlimited teenagers black socks"}
+{"eventType":"detail-page-view","visitorId":"bjbs_group1_visitor1","eventTime":"2021-12-10T10:27:42.000Z","productDetails":{"product":{"id":"GGCOGAEC100616"},"quantity":3}}
