+++ samples/interactive-tutorials/search/search-with-boost-spec.js
@@ -19,38 +19,11 @@ async function main(impersonatedPrincipal) {
   // Call Retail API to search for a products in a catalog, rerank the
   // results boosting or burying the products that match defined condition.
 
-  const { GoogleAuth, Impersonated } = require('google-auth-library');
-
-  // Acquire source credentials:
-  const auth = new GoogleAuth();
-  const client = await auth.getClient();
-  const projectId = await auth.getProjectId();
-
-  let opts = {};
-
-  if(impersonatedPrincipal) {
-    // Impersonate new credentials:
-    let targetClient = new Impersonated({
-        sourceClient: client,
-        targetPrincipal: impersonatedPrincipal,
-        lifetime: 30,
-        delegates: [],
-        targetScopes: ['https://www.googleapis.com/auth/cloud-platform']
-    });
-
-    // Instantiates a client.
-    opts = {
-        auth: {
-            projectId: projectId,
-            getClient: () => targetClient
-        }
-    };
-  } 
+  const {impersonatedAuthClient} = require('../setup/setup-cleanup');
+  const {opts, projectId} = await impersonatedAuthClient(impersonatedPrincipal);
 
   // Imports the Google Cloud client library.
   const {SearchServiceClient} = require('@google-cloud/retail');
-
-  
 
   // Instantiates a client
   const retailClient = new SearchServiceClient(opts);
@@ -61,7 +34,7 @@ async function main(impersonatedPrincipal) {
   const placement = `projects/${projectId}/locations/global/catalogs/default_catalog/placements/default_search`;
 
   // Raw search query.
-  const query = 'pen awfully';
+  const query = 'Hoodie';
 
   // A unique identifier for tracking visitors.
   const visitorId = '12345';
@@ -71,7 +44,7 @@ async function main(impersonatedPrincipal) {
     conditionBoostSpecs: [
       {
         condition: '(colorFamilies: ANY("Blue"))', // TRY OTHER CONDITIONS
-        boost: 1.0, // TRY DIFFERENT SCORES
+        boost: 0.0, // TRY DIFFERENT SCORES
       },
     ],
   };
@@ -103,14 +76,7 @@ async function main(impersonatedPrincipal) {
       autoPaginate: false,
     });
     const searchResponse = response[IResponseParams.ISearchResponse];
-    // console.log('Search result: ', JSON.stringify(searchResponse, null, 4));
-    console.log(searchResponse.results);
-    if(searchResponse) {
-        console.log("===> list results:")        
-        searchResponse.results.forEach(e => {
-            console.log(e.product.title, e.product.colorInfo);
-        });
-    }
+    console.log('Search result: ', JSON.stringify(searchResponse, null, 4));
     console.log('Search end');
   };
