+++ samples/package.json
@@ -13,7 +13,7 @@
     "test": "c8 mocha"
   },
   "dependencies": {
-    "@google-cloud/retail": "^1.8.0",
+    "@google-cloud/retail": "^1.8.1",
     "@google-cloud/bigquery": "^5.9.2",
     "@google-cloud/storage": "^5.16.1"
   },
